:truck: move BasePane class to ‘commons/Pane/state’
